improved, error-free cli parser
