+++ distage/distage-framework/src/main/scala/izumi/distage/roles/RoleAppBootModule.scala
@@ -17,7 +17,7 @@ import izumi.distage.roles.launcher.*
 import izumi.distage.roles.launcher.AppResourceProvider.{AppResource, FinalizerFilters}
 import izumi.distage.roles.launcher.ModuleValidator.ValidatedModulePair
 import izumi.distage.roles.model.meta.{LibraryReference, RolesInfo}
-import izumi.fundamentals.platform.cli.{CLIParser, CLIParserImpl, MultiModalArgsParser, MultiModalArgsParserImpl, ParserFailureHandler}
+import izumi.fundamentals.platform.cli.{CLIParser, CLIParserImpl, MultiModalArgsParser, MultiModalArgsParserImpl, ParserFailureHandler, SubArgsParser, SubArgsParserImpl}
 import izumi.fundamentals.platform.resources.IzArtifact
 import izumi.logstage.api.IzLogger
 import izumi.logstage.api.logger.LogRouter
@@ -79,6 +79,7 @@ class RoleAppBootModule[F[_]: TagK: DefaultModule](
 
   make[CLIParser].from[CLIParserImpl]
   make[MultiModalArgsParser].from[MultiModalArgsParserImpl]
+  make[SubArgsParser].from[SubArgsParserImpl]
 
   make[ParserFailureHandler].from(ParserFailureHandler.TerminatingHandler)
